chore: move get client dir to kadclient
- Updates wallet_cmds to take root dir as arg.
